extract uichanges to a file
