+++ components/navigation/content/SwapWidgetOrb.tsx
@@ -1,14 +1,13 @@
-import { useAppContext } from 'components/AppContextProvider'
 import { NavigationOrb } from 'components/navigation/NavigationMenuOrb'
 import {
   SWAP_WIDGET_CHANGE_SUBJECT,
   SwapWidgetChangeAction,
 } from 'features/swapWidget/SwapWidgetChange'
+import { uiChanges } from 'helpers/uiChanges'
 import { useOnboarding } from 'helpers/useOnboarding'
 import React, { useState } from 'react'
 
 export function SwapWidgetOrb() {
-  const { uiChanges } = useAppContext()
   const [exchangeOnboarded] = useOnboarding('SwapWidget')
   const [exchangeOpened, setExchangeOpened] = useState(false)
