+++ features/automation/api/automationTxHandlers.ts
@@ -9,7 +9,7 @@ import {
   removeAutomationBotAggregatorTriggers,
 } from 'blockchain/calls/automationBotAggregator'
 import { TransactionDef } from 'blockchain/calls/callsHelpers'
-import { AutomationTxData, TxHelpers, UIChanges } from 'components/AppContext'
+import { AutomationTxData, TxHelpers } from 'components/AppContext'
 import {
   addTransactionMap,
   AutomationAddTriggerData,
@@ -19,6 +19,7 @@ import {
 } from 'features/automation/common/txDefinitions'
 import { AutomationPublishType } from 'features/automation/common/types'
 import { handleTransaction } from 'helpers/handleTransaction'
+import { UIChanges } from 'helpers/uiChanges'
 import { takeWhileInclusive } from 'rxjs-take-while-inclusive'
 
 export const takeUntilTxState = [
