+++ features/automation/common/controls/AddAndRemoveTriggerControl.tsx
@@ -4,7 +4,6 @@ import {
 } from 'blockchain/calls/automationBotAggregator'
 import { TxMetaKind } from 'blockchain/calls/txMeta'
 import { TxHelpers } from 'components/AppContext'
-import { useAppContext } from 'components/AppContextProvider'
 import { useAutomationContext } from 'components/AutomationContextProvider'
 import { AutoBSTriggerResetData } from 'features/automation/common/state/autoBSFormChange'
 import {
@@ -21,6 +20,7 @@ import { AutomationContracts } from 'features/automation/metadata/types'
 import { AutoTakeProfitResetData } from 'features/automation/optimization/autoTakeProfit/state/autoTakeProfitFormChange'
 import { StopLossResetData } from 'features/automation/protection/stopLoss/state/StopLossFormChange'
 import { TX_DATA_CHANGE } from 'helpers/gasEstimate'
+import { uiChanges } from 'helpers/uiChanges'
 import { ReactElement, useEffect, useMemo } from 'react'
 
 export interface AddAndRemoveTxHandler {
@@ -68,7 +68,6 @@ export function AddAndRemoveTriggerControl({
   analytics,
   contracts,
 }: AddAndRemoveTriggerControlProps) {
-  const { uiChanges } = useAppContext()
   const {
     environmentData: { ethMarketPrice },
     positionData: { id, ilk, owner, positionRatio },
