+++ features/automation/common/state/useAutoBSStateInitializator.ts
@@ -1,6 +1,5 @@
 import { TriggerType } from '@oasisdex/automation'
 import BigNumber from 'bignumber.js'
-import { useAppContext } from 'components/AppContextProvider'
 import {
   prepareAutoBSSliderDefaults,
   resolveMaxBuyOrMinSellPrice,
@@ -12,6 +11,7 @@ import {
 } from 'features/automation/common/state/autoBSFormChange'
 import { AutoBSTriggerData } from 'features/automation/common/state/autoBSTriggerData'
 import { StopLossTriggerData } from 'features/automation/protection/stopLoss/state/stopLossTriggerData'
+import { uiChanges } from 'helpers/uiChanges'
 import { useEffect } from 'react'
 
 export function useAutoBSstateInitialization({
@@ -25,7 +25,6 @@ export function useAutoBSstateInitialization({
   stopLossTriggerData: StopLossTriggerData
   type: TriggerType
 }) {
-  const { uiChanges } = useAppContext()
   const {
     triggerId,
     execCollRatio,
