+++ features/automation/optimization/autoBuy/sidebars/SidebarSetupAutoBuy.tsx
@@ -1,5 +1,4 @@
 import BigNumber from 'bignumber.js'
-import { useAppContext } from 'components/AppContextProvider'
 import { useAutomationContext } from 'components/AutomationContextProvider'
 import { useGasEstimationContext } from 'components/GasEstimationContextProvider'
 import { SidebarSection, SidebarSectionProps } from 'components/sidebar/SidebarSection'
@@ -28,6 +27,7 @@ import {
   extractCancelAutomationErrors,
   extractCancelAutomationWarnings,
 } from 'helpers/messageMappers'
+import { uiChanges } from 'helpers/uiChanges'
 import React from 'react'
 import { Grid } from 'theme-ui'
 
@@ -70,7 +70,6 @@ export function SidebarSetupAutoBuy({
   isAutoBuyActive,
 }: SidebarSetupAutoBuyProps) {
   const gasEstimation = useGasEstimationContext()
-  const { uiChanges } = useAppContext()
   const {
     environmentData: { ethBalance, ethMarketPrice, etherscanUrl },
     positionData: { nextPositionRatio, token, liquidationRatio, vaultType },
