+++ features/automation/optimization/constantMultiple/controls/ConstantMultipleDetailsLayout.tsx
@@ -5,7 +5,6 @@ import {
   trackingEvents,
 } from 'analytics/analytics'
 import BigNumber from 'bignumber.js'
-import { useAppContext } from 'components/AppContextProvider'
 import { useAutomationContext } from 'components/AutomationContextProvider'
 import { Banner, bannerGradientPresets } from 'components/Banner'
 import { DetailsSection } from 'components/DetailsSection'
@@ -22,6 +21,7 @@ import {
 import { AutomationFeatures } from 'features/automation/common/types'
 import { VaultType } from 'features/generalManageVault/vaultType'
 import { EXTERNAL_LINKS } from 'helpers/applicationLinks'
+import { uiChanges } from 'helpers/uiChanges'
 import { useUIChanges } from 'helpers/uiChangesHook'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
@@ -63,7 +63,6 @@ export function ConstantMultipleDetailsLayout({
   afterTriggerColRatioToSell,
 }: ConstantMultipleDetailsLayoutProps) {
   const { t } = useTranslation()
-  const { uiChanges } = useAppContext()
   const {
     positionData: { vaultType },
   } = useAutomationContext()
