+++ features/automation/optimization/constantMultiple/sidebars/SidebarSetupConstantMultiple.tsx
@@ -1,5 +1,4 @@
 import BigNumber from 'bignumber.js'
-import { useAppContext } from 'components/AppContextProvider'
 import { useAutomationContext } from 'components/AutomationContextProvider'
 import { useGasEstimationContext } from 'components/GasEstimationContextProvider'
 import { SidebarSection, SidebarSectionProps } from 'components/sidebar/SidebarSection'
@@ -27,6 +26,7 @@ import {
   extractCancelAutomationErrors,
   extractCancelAutomationWarnings,
 } from 'helpers/messageMappers'
+import { uiChanges } from 'helpers/uiChanges'
 import React from 'react'
 import { Grid } from 'theme-ui'
 
@@ -78,7 +78,6 @@ export function SidebarSetupConstantMultiple({
   debtDeltaAfterSell,
 }: SidebarSetupConstantMultipleProps) {
   const gasEstimation = useGasEstimationContext()
-  const { uiChanges } = useAppContext()
   const {
     environmentData: { ethBalance, ethMarketPrice, etherscanUrl },
     positionData: { debt, debtFloor, nextPositionRatio, token, vaultType },
