+++ features/navigation/controls/AjnaNavigationController.tsx
@@ -1,4 +1,3 @@
-import { useAppContext } from 'components/AppContextProvider'
 import { MyPositionsLink } from 'components/navigation/content/MyPositionsLink'
 import { Navigation, navigationBreakpoints } from 'components/navigation/Navigation'
 import { SwapWidgetShowHide } from 'components/swapWidget/SwapWidgetShowHide'
@@ -8,12 +7,12 @@ import {
   SwapWidgetChangeAction,
 } from 'features/swapWidget/SwapWidgetChange'
 import { INTERNAL_LINKS } from 'helpers/applicationLinks'
+import { uiChanges } from 'helpers/uiChanges'
 import { useAccount } from 'helpers/useAccount'
 import React from 'react'
 import { useMediaQuery } from 'usehooks-ts'
 
 export function AjnaNavigationController() {
-  const { uiChanges } = useAppContext()
   const { isConnected, walletAddress } = useAccount()
   const isViewBelowXl = useMediaQuery(`(max-width: ${navigationBreakpoints[3] - 1}px)`)
