+++ helpers/uiChangesHook.tsx
@@ -1,10 +1,8 @@
-import { SupportedUIChangeType } from 'components/AppContext'
-import { useAppContext } from 'components/AppContextProvider'
 import { useEffect, useState } from 'react'
 
-export function useUIChanges<T extends SupportedUIChangeType>(topic: string): T[] {
-  const { uiChanges } = useAppContext()
+import { SupportedUIChangeType, uiChanges } from './uiChanges'
 
+export function useUIChanges<T extends SupportedUIChangeType>(topic: string): T[] {
   const [lastUIState, lastUIStateSetter] = useState(uiChanges.lastPayload<T>(topic))
 
   useEffect(() => {
